requirements.txt
---
title: requirements.txt
date: 2025-10-16
version: draft
checksum: c0bcd9665686
---

# HYMetaLab Open-Source Data Integration - Core Dependencies
# Version: 1.0
# Date: 2025-10-15

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Scientific Computing
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & Serialization
pyyaml>=6.0

# Optional: Web Requests (for live data APIs)
requests>=2.31.0

# Optional: Progress Bars
tqdm>=4.65.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality (optional, for development)
black>=23.0.0
flake8>=6.0.0


## Methods
Briefly state datasets, parameters, seeds, and procedures.

## Limitations
List key caveats (sampling bias, small N, model assumptions).

## Evidence & Links
- [Link 1](#)
- [Link 2](#)

Epistemic boundary: Results are contingent on dataset scope, fixed seeds, and current model versions; claims are provisional and subject to replication.
